backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
psycopg2-binary==2.9.10
google-genai==1.29.0
httpx==0.28.1
SQLAlchemy==2.0.36
